Refactor asynchia.ee to remove InputEOF and CollectorFull exceptions.
Prior to this changeset, asynchia.ee was prone to a very subtle issue.
If there was a DelimitedCollector that collected 5 bytes,
its onclose function was not called if only 5 bytes were sent over
the connection of all. To resolve this issue, it is now possible for
collectors and inputs to receive/send data and signal that they are
full at the same time. It also seems to have simplified the code
quite a bit.